+++ lib/view/admin/alunos.view.dart
@@ -1,8 +1,8 @@
-import 'package:apoio_forca_treino/view/aluno.item.dart';
+import 'package:apoio_forca_treino/view/admin/aluno.item.dart';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 
-import '../model/aluno.model.dart';
+import '../../model/aluno.model.dart';
 
 class AlunoView extends StatelessWidget {
   FirebaseFirestore firestore = FirebaseFirestore.instance;
