+++ lib/view/app.dart
@@ -4,16 +4,16 @@ import 'package:flutter/material.dart';
 import 'login.view.dart';
 import 'alteracaoSenha.view.dart';
 
-import 'home.admin.view.dart';
-import 'homeprofessor.dart';
-import 'home.aluno.view.dart';
+import 'admin/home.admin.view.dart';
+import 'professor/homeprofessor.dart';
+import 'alunos/home.aluno.view.dart';
 
-import 'cad.alunos.view.dart';
-import 'cad.professor.view.dart';
-import 'cad.exercicio.view.dart';
+import 'admin/cad.alunos.view.dart';
+import 'admin/cad.professor.view.dart';
+import 'professor/cad.exercicio.view.dart';
 
-import 'alunos.view.dart';
-import 'prof.view.dart';
+import 'admin/alunos.view.dart';
+import 'admin/prof.view.dart';
 
 class App extends StatelessWidget {
   final auth = FirebaseAuth.instance;
